인벤토리 시스템 업그레이드중
업그레이드..

+++ JsonFile/Assets/Json2/Weapon_Master.cs
@@ -9,7 +9,7 @@ public class Weapon_Master
     public int Weapon_DMG;
     public float STR_Scaling;
     public float DEX_Scaling;
-    public int INT_Scaling;
+    public float INT_Scaling;
     public float MAG_Scaling;
     public int DIV_Scaling;
     public int CHR_Scaling;
